chore(aligned): create the `ptr` module
