+++ bootloader/uefi_wrapper/src/system_table.rs
@@ -1,5 +1,6 @@
 use crate::protocols::console;
 use crate::service;
+use aligned::ptr;
 use core::fmt;
 use r_efi::efi;
 
@@ -16,7 +17,7 @@ impl SystemTable {
         // A value of `SystemTable` is created only through the argument of `efi_main`. Since this method
         // takes a mutable reference and this type does not implement `Copy`, only one mutable
         // reference to `efi::BootServices` is created.
-        let bs = unsafe { aligned::as_mut(st.boot_services) };
+        let bs = unsafe { ptr::as_mut(st.boot_services) };
 
         service::Boot::new(bs, self)
     }
@@ -29,7 +30,7 @@ impl SystemTable {
         // `SystemTable` is created only through the argument of `efi_main`. Since this method
         // takes a mutable reference and this type does not implement `Copy`, only one mutable
         // reference to `efi::simple_text_output::Protocol` is created.
-        let con_out = unsafe { aligned::as_mut(st.con_out) };
+        let con_out = unsafe { ptr::as_mut(st.con_out) };
 
         console::SimpleTextOutput::new(con_out, self)
     }
@@ -39,7 +40,7 @@ impl SystemTable {
         //
         // A value of `SystemTable` is created only through the argument of `efi_main`. Since this method takes a mutable
         // reference of an instance and this type does not implement `Copy`, only one mutable reference to `efi::SystemTable` is created.
-        unsafe { aligned::as_mut(self.0) }
+        unsafe { ptr::as_mut(self.0) }
     }
 }
 impl fmt::Debug for SystemTable {
